removed useless FRand wrapper, and small things

+++ src/org/recast4j/recast/RecastCompact.java
@@ -82,7 +82,7 @@ public class RecastCompact {
                 while (s != null) {
                     if (s.area != RC_NULL_AREA) {
                         int bot = s.smax;
-                        int top = s.next != null ? (int) s.next.smin : MAX_HEIGHT;
+                        int top = s.next != null ? s.next.smin : MAX_HEIGHT;
                         chf.spans[idx].y = RecastCommon.clamp(bot, 0, MAX_HEIGHT);
                         chf.spans[idx].h = RecastCommon.clamp(top - bot, 0, MAX_HEIGHT);
                         chf.areas[idx] = s.area;
